Enabled buying products using 'buy now' with cash-on delivery payment method

+++ src/Entity/Purchase.php
@@ -27,7 +27,7 @@ class Purchase
     private $user;
 
     /**
-     * @ORM\Column(type="integer")
+     * @ORM\Column(type="float")
      */
     private $price;
 
@@ -63,12 +63,12 @@ class Purchase
         return $this;
     }
 
-    public function getPrice(): ?int
+    public function getPrice(): ?float
     {
         return $this->price;
     }
 
-    public function setPrice(int $price): self
+    public function setPrice(float $price): self
     {
         $this->price = $price;
